pyproject.toml
[project]
name = "simple-database-connection-application"
version = "0.1"
dependencies = [
    "psycopg2-binary==2.9.9"
]
requires-python = ">=3.9,<4"
readme = "README.md"
